Fixed DasServlet. Had left out import statement and addition of .getService(HttpServletRequest) method.
"How?" you ask? Well, I did my prior edit in notepad rather than in Eclipse, because Eclipse would reformat the source code according to my preferences, which would be a major change to the file instead of just small adjustments. Of course this would be fixed if IBM would store the formatting preferences in the project itself so that all developers would have the same rules applied by their IDE.