Update JsonJdbcQueryContent.java
When returning a jdbc query via json through the Jdbcqueryrest service, there is no check for null when writing the column values.
This means if any column values are null it will throw a null pointer exception and get an Error 500.
This can be handled by checking for null and outputting null , or alternatively you can just ignore it I guess? in any case it should try to avoid the null pointer exception.