Bugfix so that you can clear a single Cache entry in userBean / peopleBean / any AbstractResourceProvider
Due to incorrect logic it was impossible to clear the cache / refresh for a single entry in the cache.

For example when using the userBean / peopleBean, if you tried to use peopleBean.clearCache(@UserName) it had no effect.

This is because of 3 bugs in the clearCache method

1. the tests for each scope if (c==null) should actually be if (c!=null). This was preventing the code from entering into the 'remove' condition
2. The 'remove' method was incorrectly being called on the scope map e.g. sessionScope instead of the system cache that was retrieved
3. The 'remove' method was using the 'CACHE_KEY' when it should actually be using the id

Note: the clearCache() method to clear the entire cache was working correctly, however I think this must have been copy/pasted to the clearCache(String) method and not correctly tidied up